SSOR-61 Refactored to use boss-core. Added swagger.
