+++ src/main/java/org/ssor/boss/card/BossCardServiceApplication.java
@@ -2,8 +2,11 @@ package org.ssor.boss.card;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @SpringBootApplication
+@ComponentScan("org.ssor.boss.core.repository")
 public class BossCardServiceApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(BossCardServiceApplication.class, args);
