+++ src/main/resources/application.properties
@@ -1,7 +1,5 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/boss
 spring.datasource.username=root
 spring.datasource.password=password
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-server.port=8090
-#spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
